+++ linen-repository/src/main/scala/x7c1/linen/repository/preset/PresetFactory.scala
@@ -1,8 +1,8 @@
 package x7c1.linen.repository.preset
 
 import x7c1.linen.database.LinenOpenHelper
-import x7c1.linen.repository.account.preset.{PresetAccountSetup, PresetChannelSetup}
-import x7c1.linen.repository.channel.preset.PresetChannelPiece
+import x7c1.linen.repository.account.setup.PresetAccountSetup
+import x7c1.linen.repository.channel.preset.{PresetChannelSetup, PresetChannelPiece}
 import x7c1.linen.repository.source.setting.{ChannelOwner, ChannelSourceParts}
 import x7c1.wheat.macros.logger.Log
